% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{makeObject}
\alias{makeObject}
\title{Function that creates downloads data and creates SummarizedExperiment object}
\usage{
makeObject(identifier, identifier_list = identifiers)
}
\arguments{
\item{identifier}{name of data set}

\item{identifier_list}{list of identifiers, defaults to current list}
}
\value{
a SummarizedExperiment object for the data set
}
\description{
Takes a name of data set and passes to identifiers named list
Uses the vector of identifiers in the list to get URL and download
}
\examples{
makeObject('GSE41197')
}
